backend_admin_python/requirements.txt
# FastAPI核心依赖
fastapi==0.110.0
uvicorn[standard]==0.29.0
pydantic==2.7.0
pydantic-settings==2.2.0

# 数据库相关
sqlalchemy==2.0.30
pymysql==1.1.0
cryptography==42.0.7
alembic==1.13.1

# 认证相关
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9

# 报告生成相关
matplotlib==3.8.4
pandas==2.2.2
numpy==1.26.4
reportlab==3.6.12  # PDF报告生成库

# 工具类
python-dotenv==1.0.1
aiofiles==23.2.1
flask==3.0.3

# 测试相关
pytest==8.2.0
pytest-asyncio==0.23.6
httpx==0.27.0
